+++ darkside-test-lib/src/androidTest/java/cash/z/ecc/android/sdk/darkside/test/TestWallet.kt
@@ -97,12 +97,8 @@ class TestWallet(
                 throw TimeoutException("Failed to sync wallet within ${timeout}ms")
             }
         }
-        if (!synchronizer.isStarted) {
-            twig("Starting sync")
-            synchronizer.start(walletScope)
-        } else {
-            twig("Awaiting next SYNCED status")
-        }
+
+        twig("Awaiting next SYNCED status")
 
         // block until synced
         synchronizer.status.first { it == Synchronizer.Status.SYNCED }
@@ -154,7 +150,7 @@ class TestWallet(
             twig("Scheduling a stop in ${timeout}ms")
             walletScope.launch {
                 delay(timeout)
-                synchronizer.stop()
+                synchronizer.close()
             }
         }
         synchronizer.status.first { it == Synchronizer.Status.STOPPED }