+++ demo-app/src/main/java/cash/z/ecc/android/sdk/demoapp/demos/listutxos/ListUtxosFragment.kt
@@ -4,10 +4,10 @@ import android.content.Context
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
+import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
+import androidx.lifecycle.repeatOnLifecycle
 import androidx.recyclerview.widget.LinearLayoutManager
-import cash.z.ecc.android.bip39.Mnemonics
-import cash.z.ecc.android.bip39.toSeed
 import cash.z.ecc.android.sdk.SdkSynchronizer
 import cash.z.ecc.android.sdk.Synchronizer
 import cash.z.ecc.android.sdk.block.CompactBlockProcessor
@@ -16,17 +16,16 @@ import cash.z.ecc.android.sdk.demoapp.databinding.FragmentListUtxosBinding
 import cash.z.ecc.android.sdk.demoapp.ext.requireApplicationContext
 import cash.z.ecc.android.sdk.demoapp.util.fromResources
 import cash.z.ecc.android.sdk.demoapp.util.mainActivity
-import cash.z.ecc.android.sdk.ext.collectWith
 import cash.z.ecc.android.sdk.internal.twig
 import cash.z.ecc.android.sdk.model.Account
 import cash.z.ecc.android.sdk.model.BlockHeight
-import cash.z.ecc.android.sdk.model.LightWalletEndpoint
 import cash.z.ecc.android.sdk.model.TransactionOverview
 import cash.z.ecc.android.sdk.model.ZcashNetwork
-import cash.z.ecc.android.sdk.model.defaultForNetwork
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.delay
-import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.flow.filterNotNull
+import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import kotlin.math.max
@@ -45,8 +44,6 @@ import kotlin.math.max
  */
 @Suppress("TooManyFunctions")
 class ListUtxosFragment : BaseDemoFragment<FragmentListUtxosBinding>() {
-    private lateinit var seed: ByteArray
-    private lateinit var synchronizer: Synchronizer
     private lateinit var adapter: UtxoAdapter
     private val address: String = "t1RwbKka1CnktvAJ1cSqdn7c6PXWG4tZqgd"
     private var status: Synchronizer.Status? = null
@@ -55,30 +52,6 @@ class ListUtxosFragment : BaseDemoFragment<FragmentListUtxosBinding>() {
 
     override fun inflateBinding(layoutInflater: LayoutInflater): FragmentListUtxosBinding =
         FragmentListUtxosBinding.inflate(layoutInflater)
-
-    /**
-     * Initialize the required values that would normally live outside the demo but are repeated
-     * here for completeness so that each demo file can serve as a standalone example.
-     */
-    private fun setup() {
-        // Use a BIP-39 library to convert a seed phrase into a byte array. Most wallets already
-        // have the seed stored
-        seed = Mnemonics.MnemonicCode(sharedViewModel.seedPhrase.value).toSeed()
-        val network = ZcashNetwork.fromResources(requireApplicationContext())
-        synchronizer = Synchronizer.newBlocking(
-            requireApplicationContext(),
-            network,
-            alias = "Demo_Utxos",
-            lightWalletEndpoint = LightWalletEndpoint.defaultForNetwork(network),
-            seed = seed,
-            birthday = sharedViewModel.birthdayHeight.value
-        )
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setup()
-    }
 
     private fun initUi() {
         binding.inputAddress.setText(address)
@@ -96,57 +69,59 @@ class ListUtxosFragment : BaseDemoFragment<FragmentListUtxosBinding>() {
     }
 
     private fun downloadTransactions() {
-        binding.textStatus.text = "loading..."
-        binding.textStatus.post {
-            val network = ZcashNetwork.fromResources(requireApplicationContext())
-            binding.textStatus.requestFocus()
-            val addressToUse = binding.inputAddress.text.toString()
-            val startToUse = max(
-                binding.inputRangeStart.text.toString().toLongOrNull()
-                    ?: network.saplingActivationHeight.value,
-                network.saplingActivationHeight.value
-            )
-            val endToUse = binding.inputRangeEnd.text.toString().toLongOrNull()
-                ?: getUxtoEndHeight(requireApplicationContext()).value
-            var allStart = now
-            twig("loading transactions in range $startToUse..$endToUse")
-            val txids = lightWalletService?.getTAddressTransactions(
-                addressToUse,
-                BlockHeight.new(network, startToUse)..BlockHeight.new(network, endToUse)
-            )
-            var delta = now - allStart
-            updateStatus("found ${txids?.size} transactions in ${delta}ms.", false)
+        sharedViewModel.synchronizerFlow.value?.let { synchronizer ->
+            binding.textStatus.text = "loading..."
+            binding.textStatus.post {
+                val network = ZcashNetwork.fromResources(requireApplicationContext())
+                binding.textStatus.requestFocus()
+                val addressToUse = binding.inputAddress.text.toString()
+                val startToUse = max(
+                    binding.inputRangeStart.text.toString().toLongOrNull()
+                        ?: network.saplingActivationHeight.value,
+                    network.saplingActivationHeight.value
+                )
+                val endToUse = binding.inputRangeEnd.text.toString().toLongOrNull()
+                    ?: getUxtoEndHeight(requireApplicationContext()).value
+                var allStart = now
+                twig("loading transactions in range $startToUse..$endToUse")
+                val txids = lightWalletService?.getTAddressTransactions(
+                    addressToUse,
+                    BlockHeight.new(network, startToUse)..BlockHeight.new(network, endToUse)
+                )
+                var delta = now - allStart
+                updateStatus("found ${txids?.size} transactions in ${delta}ms.", false)
 
-            txids?.map {
-                // Disabled during migration to newer SDK version; this appears to have been
-                // leveraging non-public  APIs in the SDK so perhaps should be removed
-                // it.data.apply {
-                //     try {
-                //         runBlocking { initializer.rustBackend.decryptAndStoreTransaction(toByteArray()) }
-                //     } catch (t: Throwable) {
-                //         twig("failed to decrypt and store transaction due to: $t")
-                //     }
-                // }
-            }?.let { _ ->
-                // Disabled during migration to newer SDK version; this appears to have been
-                // leveraging non-public  APIs in the SDK so perhaps should be removed
-                // val parseStart = now
-                // val tList = LocalRpcTypes.TransactionDataList.newBuilder().addAllData(txData).build()
-                // val parsedTransactions = initializer.rustBackend.parseTransactionDataList(tList)
-                // delta = now - parseStart
-                // updateStatus("parsed txs in ${delta}ms.")
-            }
-            (synchronizer as SdkSynchronizer).refreshTransactions()
-            delta = now - allStart
-            updateStatus("Total time ${delta}ms.")
+                txids?.map {
+                    // Disabled during migration to newer SDK version; this appears to have been
+                    // leveraging non-public  APIs in the SDK so perhaps should be removed
+                    // it.data.apply {
+                    //     try {
+                    //         runBlocking { initializer.rustBackend.decryptAndStoreTransaction(toByteArray()) }
+                    //     } catch (t: Throwable) {
+                    //         twig("failed to decrypt and store transaction due to: $t")
+                    //     }
+                    // }
+                }?.let { _ ->
+                    // Disabled during migration to newer SDK version; this appears to have been
+                    // leveraging non-public  APIs in the SDK so perhaps should be removed
+                    // val parseStart = now
+                    // val tList = LocalRpcTypes.TransactionDataList.newBuilder().addAllData(txData).build()
+                    // val parsedTransactions = initializer.rustBackend.parseTransactionDataList(tList)
+                    // delta = now - parseStart
+                    // updateStatus("parsed txs in ${delta}ms.")
+                }
+                (synchronizer as SdkSynchronizer).refreshTransactions()
+                delta = now - allStart
+                updateStatus("Total time ${delta}ms.")
 
-            lifecycleScope.launch {
-                withContext(Dispatchers.IO) {
-                    finalCount = (synchronizer as SdkSynchronizer).getTransactionCount().toInt()
-                    withContext(Dispatchers.Main) {
-                        @Suppress("MagicNumber")
-                        delay(100)
-                        updateStatus("Also found ${finalCount - initialCount} shielded txs")
+                lifecycleScope.launch {
+                    withContext(Dispatchers.IO) {
+                        finalCount = synchronizer.getTransactionCount()
+                        withContext(Dispatchers.Main) {
+                            @Suppress("MagicNumber")
+                            delay(100)
+                            updateStatus("Also found ${finalCount - initialCount} shielded txs")
+                        }
                     }
                 }
             }
@@ -167,72 +142,58 @@ class ListUtxosFragment : BaseDemoFragment<FragmentListUtxosBinding>() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         initUi()
-    }
-
-    override fun onResume() {
-        super.onResume()
-        resetInBackground()
-        viewLifecycleOwner.lifecycleScope.launchWhenStarted {
-            binding.inputAddress.setText(
-                synchronizer.getTransparentAddress(Account.DEFAULT)
-            )
-        }
+        monitorStatus()
     }
 
     var initialCount: Int = 0
     var finalCount: Int = 0
-
-    @Suppress("TooGenericExceptionCaught")
-    private fun resetInBackground() {
-        try {
-            lifecycleScope.launch {
-                withContext(Dispatchers.IO) {
-                    synchronizer.prepare()
-                    initialCount = (synchronizer as SdkSynchronizer).getTransactionCount().toInt()
-                }
-
-                onTransactionsUpdated(synchronizer.clearedTransactions.first())
-            }
-
-            // synchronizer.receivedTransactions.collectWith(lifecycleScope, ::onTransactionsUpdated)
-        } catch (t: Throwable) {
-            twig("failed to start the synchronizer!!! due to : $t")
-        }
-    }
-
-    fun onResetComplete() {
-        initTransactionUi()
-        startSynchronizer()
-        monitorStatus()
-    }
-
-    fun onClear() {
-        synchronizer.stop()
-    }
 
     private fun initTransactionUi() {
         binding.recyclerTransactions.layoutManager =
             LinearLayoutManager(activity, LinearLayoutManager.VERTICAL, false)
         adapter = UtxoAdapter()
         binding.recyclerTransactions.adapter = adapter
-        // lifecycleScope.launch {
-        //     address = synchronizer.getAddress()
-        //     synchronizer.receivedTransactions.onEach {
-        //         onTransactionsUpdated(it)
-        //     }.launchIn(this)
-        // }
     }
 
-    private fun startSynchronizer() {
-        lifecycleScope.apply {
-            synchronizer.start(this)
-        }
-    }
-
+    @OptIn(ExperimentalCoroutinesApi::class)
     private fun monitorStatus() {
-        synchronizer.status.collectWith(lifecycleScope, ::onStatus)
-        synchronizer.processorInfo.collectWith(lifecycleScope, ::onProcessorInfoUpdated)
-        synchronizer.progress.collectWith(lifecycleScope, ::onProgress)
+        viewLifecycleOwner.lifecycleScope.launch {
+            viewLifecycleOwner.repeatOnLifecycle(Lifecycle.State.STARTED) {
+                launch {
+                    sharedViewModel.synchronizerFlow
+                        .filterNotNull()
+                        .flatMapLatest { it.status }
+                        .collect { onStatus(it) }
+                }
+                launch {
+                    sharedViewModel.synchronizerFlow
+                        .filterNotNull()
+                        .flatMapLatest { it.progress }
+                        .collect { onProgress(it) }
+                }
+                launch {
+                    sharedViewModel.synchronizerFlow
+                        .filterNotNull()
+                        .flatMapLatest { it.processorInfo }
+                        .collect { onProcessorInfoUpdated(it) }
+                }
+                launch {
+                    sharedViewModel.synchronizerFlow
+                        .filterNotNull()
+                        .flatMapLatest { it.clearedTransactions }
+                        .collect { onTransactionsUpdated(it) }
+                }
+                launch {
+                    sharedViewModel.synchronizerFlow
+                        .filterNotNull()
+                        .collect {
+                            binding.inputAddress.setText(
+                                it.getTransparentAddress(Account.DEFAULT)
+                            )
+                        }
+                }
+            }
+        }
     }
 
     private fun onProcessorInfoUpdated(info: CompactBlockProcessor.ProcessorInfo) {
@@ -262,10 +223,12 @@ class ListUtxosFragment : BaseDemoFragment<FragmentListUtxosBinding>() {
     override fun onActionButtonClicked() {
         lifecycleScope.launch {
             withContext(Dispatchers.IO) {
-                twig("current count: ${(synchronizer as SdkSynchronizer).getTransactionCount()}")
-                twig("refreshing transactions")
-                (synchronizer as SdkSynchronizer).refreshTransactions()
-                twig("current count: ${(synchronizer as SdkSynchronizer).getTransactionCount()}")
+                sharedViewModel.synchronizerFlow.value?.let { synchronizer ->
+                    twig("current count: ${(synchronizer as SdkSynchronizer).getTransactionCount()}")
+                    twig("refreshing transactions")
+                    synchronizer.refreshTransactions()
+                    twig("current count: ${synchronizer.getTransactionCount()}")
+                }
             }
         }
     }